Refactor database schema and update model classes
- Renamed `Roles.sql` to `Role.sql` and updated related SQL files to reflect new data types and constraints.
- Changed primary keys in several tables and updated corresponding model classes in C#.
- Removed `CourseDetails.cs` and created new classes for `ExamMonth`, `UserQPTemplateTag`, and others.
- Updated `ExaminationDbContext` to reflect new class names and properties.
- Adjusted property names in various service and controller files for consistency.
- Modified connection string in `appsettings.json` to use `localhost`.

+++ Database/SECE.Database/Tables/DegreeType.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[DegreeType]
 (
-	[DegreeTypeId] INT NOT NULL PRIMARY KEY,
+	[DegreeTypeId] BIGINT NOT NULL PRIMARY KEY,
     [Name] NVARCHAR(100) NOT NULL,
     [Code] NVARCHAR(100) NULL,
     IsActive BIT DEFAULT 1,
